create first microservice architecture
