+++ README.md
@@ -9,6 +9,7 @@
 - Separate controllers depending on REST conventions
 - Make Tracker implements Runnable and use ScheduledExecutorService
 - Implement CompletableFuture and Executor with fixed thread pool
+- Create first microservice structure with BuildSrc
 
 ###3 - Migrate Gradle from Groovy to Kotlin Script
 - Adapt syntax to look like Kts
