+++ applications/dummy/src/main/java/tourGuide/service/TourGuideService.java
@@ -1,0 +1,39 @@
+package tourGuide.service;
+
+import gpsUtil.location.Attraction;
+import gpsUtil.location.VisitedLocation;
+import org.openclassroom.tourguide.model.User;
+import org.openclassroom.tourguide.model.UserReward;
+import tripPricer.Provider;
+
+import java.util.List;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutorService;
+
+public interface TourGuideService {
+
+    List<UserReward> getUserRewards(User user);
+
+    VisitedLocation getUserLocation(User user);
+
+    User getUser(String userName);
+
+    List<User> getAllUsers();
+
+    void addUser(User user);
+
+    List<Provider> getTripDeals(User user);
+
+    CompletableFuture<?> trackUserLocation(User user);
+
+    List<Attraction> getNearByAttractions(VisitedLocation visitedLocation);
+
+    ExecutorService getExecutor();
+
+    void stopTracker();
+
+    void clearVisitedLocationsOfUser(User user);
+
+    VisitedLocation addToVisitedLocationsOfUser(VisitedLocation visitedLocation, User user);
+
+}
