+++ applications/dummy/src/main/resources/application-test.properties
