+++ applications/dummy/src/main/resources/application.properties
