+++ buildSrc/src/main/kotlin/DependenciesVersions.kt
@@ -1,0 +1,5 @@
+object DependenciesVersions {
+    const val shadowPlugin = "7.0.0"
+    const val springDependencyManagement = "1.0.11.RELEASE"
+    const val springBoot = "2.4.5"
+}
