+++ core-librairies/gps-utils/libs/gpsUtil.jar
