+++ settings.gradle.kts
@@ -1,12 +1,23 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * The settings file is used to specify which projects to include in your build.
- * 
- * Detailed information about configuring a multi-project build in Gradle can be found
- * in the user guide at https://docs.gradle.org/4.8.1/userguide/multi_project_builds.html
- *
- * NB : migrate to Gradle 7.0.2
- */
+subdir("applications") {
+    include("dummy")
+}
 
-rootProject.name = "TourGuide"
+subdir("core-librairies") {
+    include("gps-utils")
+    include("reward-central")
+    include("trip-pricer")
+    include("models")
+}
+
+class IncludeDsl(val root: String) {
+    fun include(project: String) {
+        settings.include(project)
+        settings.project(":$project").also {
+            it.projectDir = file("$root/$project")
+        }
+    }
+}
+
+fun subdir(root: String, block: IncludeDsl.() -> Unit) {
+    block(IncludeDsl(root))
+}
