--- src/main/java/tourGuide/model/UserReward.java
@@ -1,40 +1,0 @@
-package tourGuide.model;
-
-import gpsUtil.location.Attraction;
-import gpsUtil.location.VisitedLocation;
-
-public class UserReward {
-
-	private final VisitedLocation visitedLocation;
-	private final Attraction attraction;
-	private int rewardPoints;
-
-	public UserReward(VisitedLocation visitedLocation, Attraction attraction, int rewardPoints) {
-		this.visitedLocation = visitedLocation;
-		this.attraction = attraction;
-		this.rewardPoints = rewardPoints;
-	}
-
-	//TODO : pourquoi ?
-	public UserReward(VisitedLocation visitedLocation, Attraction attraction) {
-		this.visitedLocation = visitedLocation;
-		this.attraction = attraction;
-	}
-
-	public VisitedLocation getVisitedLocation() {
-		return visitedLocation;
-	}
-
-	public Attraction getAttraction() {
-		return attraction;
-	}
-
-	public int getRewardPoints() {
-		return rewardPoints;
-	}
-
-	public void setRewardPoints(final int rewardPoints1) {
-		rewardPoints = rewardPoints1;
-	}
-	
-}
