--- src/test/java/tourGuide/TestRewardsServiceImpl.java
@@ -1,82 +1,0 @@
-package tourGuide;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import java.util.Date;
-import java.util.List;
-import java.util.Locale;
-import java.util.UUID;
-
-import gpsUtil.GpsUtil;
-import gpsUtil.location.Attraction;
-import gpsUtil.location.VisitedLocation;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-import org.springframework.test.context.ActiveProfiles;
-import rewardCentral.RewardCentral;
-import tourGuide.helper.InternalTestHelper;
-import tourGuide.service.RewardsService;
-import tourGuide.service.RewardsServiceImpl;
-import tourGuide.service.TourGuideService;
-import tourGuide.service.TourGuideServiceImpl;
-import tourGuide.model.User;
-import tourGuide.model.UserReward;
-
-@ActiveProfiles("test")
-public class TestRewardsServiceImpl {
-
-	private RewardsService rewardsService;
-	private TourGuideService tourGuideService;
-	private GpsUtil gpsUtil;
-	private RewardCentral rewardCentral;
-
-	@BeforeAll
-	public static void setUp() {
-		Locale.setDefault(Locale.US);
-	}
-
-	@BeforeEach
-	public void beforeEach() {
-		gpsUtil = new GpsUtil();
-		rewardCentral = new RewardCentral();
-		rewardsService = new RewardsServiceImpl(gpsUtil, rewardCentral);
-		//tourGuideService = new TourGuideServiceImpl(gpsUtil, rewardsService);
-	}
-
-	@Test
-	public void userGetRewards() {
-		InternalTestHelper.setInternalUserNumber(0);
-		tourGuideService = new TourGuideServiceImpl(gpsUtil, rewardsService);
-		
-		User user = new User(UUID.randomUUID(), "jon", "000", "jon@tourGuide.com");
-		Attraction attraction = gpsUtil.getAttractions().get(0);
-		tourGuideService.addToVisitedLocationsOfUser(new VisitedLocation(user.getUserId(), attraction, new Date()), user);
-		tourGuideService.trackUserLocation(user).isDone();
-
-		List<UserReward> userRewards = user.getUserRewards();
-		assertEquals(1, userRewards.size());
-	}
-	
-	@Test
-	public void isWithinAttractionProximity() {
-		Attraction attraction = gpsUtil.getAttractions().get(0);
-		assertTrue(rewardsService.isWithinAttractionProximity(attraction, attraction));
-	}
-	
-	@Test
-	public void nearAllAttractions() {
-		rewardsService.setProximityBuffer(Integer.MAX_VALUE);
-
-		InternalTestHelper.setInternalUserNumber(1);
-		tourGuideService = new TourGuideServiceImpl(gpsUtil, rewardsService);
-		
-		rewardsService.calculateRewards(tourGuideService.getAllUsers().get(0)).join();
-		List<UserReward> userRewards = tourGuideService.getUserRewards(tourGuideService.getAllUsers().get(0));
-
-		assertEquals(gpsUtil.getAttractions().size(), userRewards.size());
-	}
-	
-}
